ingestion/requirements.txt
requests==2.31.0
google-cloud-bigquery==3.17.2
